requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
numpy==1.26.4
opencv-python==4.8.1.78
pymupdf==1.23.25
python-pptx==0.6.22
pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.5.3
python-jose==3.3.0
aiofiles==23.2.1 